+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - TBD
+### Added
+- Old locations for Table and Partitions to ApiarySNSListener events.
+- Table parameters with a configurable regular expression to control which table parameters will be added to the ApiarySNSListener message. Default will be none.
+
 ## [1.1.0] - 2019-02-15
 ### Added
 - Apiary ReadOnly Auth Event Listener.